requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.1
